Move sample out to ImageSources section
