+++ samples/CommunityToolkit.Maui.Sample/ViewModels/ImageSources/ImageSourcesGalleryViewModel.cs
@@ -1,0 +1,16 @@
+﻿using CommunityToolkit.Maui.Converters;
+using CommunityToolkit.Maui.Sample.Models;
+
+namespace CommunityToolkit.Maui.Sample.ViewModels.ImageSources;
+
+public class ImageSourcesGalleryViewModel : BaseGalleryViewModel
+{
+	public ImageSourcesGalleryViewModel()
+		: base(new[]
+		{
+			SectionModel.Create<GravatarImageSourceViewModel>(nameof(GravatarImageSource),
+				"GravatarImageSource allows you to use as an Image source, a users Gravatar registered image via their email address."),
+		})
+	{
+	}
+}
